final version after tests

+++ src/v2/core/domain/PceModule.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface PceModule extends Module {
 
     ChannelInfo heartbeat();
-    byte allocateAddress(long sid, byte mediatorId, double mediatorRetx);
-    CorrespondenceRegister correspondence(byte address);
+    int allocateAddress(long sid, byte mediatorId, double mediatorRetx);
+    CorrespondenceRegister correspondence(int address);
     List<String> feed(long controllerId, Message message);
 
 }
